chore: move cross validation into SupervisorClient (#15374)
