+++ crates/optimism/txpool/src/validator.rs
@@ -1,6 +1,6 @@
 use crate::{
     interop::{MaybeInteropTransaction, TransactionInterop},
-    supervisor::{is_valid_cross_tx, SupervisorClient},
+    supervisor::SupervisorClient,
     InvalidCrossTx,
 };
 use alloy_consensus::{BlockHeader, Transaction};
@@ -298,19 +298,20 @@ where
         outcome
     }
 
-    /// Wrapper for [`is_valid_cross_tx`]
+    /// Wrapper for is valid cross tx
     pub async fn is_valid_cross_tx(&self, tx: &Tx) -> Option<Result<(), InvalidCrossTx>> {
         // We don't need to check for deposit transaction in here, because they won't come from
         // txpool
-        is_valid_cross_tx(
-            tx.access_list(),
-            tx.hash(),
-            self.block_info.timestamp.load(Ordering::Relaxed),
-            Some(TRANSACTION_VALIDITY_WINDOW_SECS),
-            self.fork_tracker.is_interop_activated(),
-            self.supervisor_client.as_ref(),
-        )
-        .await
+        self.supervisor_client
+            .as_ref()?
+            .is_valid_cross_tx(
+                tx.access_list(),
+                tx.hash(),
+                self.block_info.timestamp.load(Ordering::Relaxed),
+                Some(TRANSACTION_VALIDITY_WINDOW_SECS),
+                self.fork_tracker.is_interop_activated(),
+            )
+            .await
     }
 }
